Based on end date and current date the commodity will be displayed and context api changes
